+++ frontend/src/metabase/query_builder/components/view/ViewHeader/components/AdHocQuestionLeftSide/AdHocQuestionLeftSide.jsx
@@ -1,0 +1,72 @@
+import PropTypes from "prop-types";
+import { t } from "ttag";
+
+import CS from "metabase/css/core/index.css";
+import {
+  AdHocLeftSideRoot,
+  AdHocViewHeading,
+  ViewHeaderLeftSubHeading,
+  ViewHeaderMainLeftContentContainer,
+} from "metabase/query_builder/components/view/ViewHeader/ViewHeader.styled";
+import {
+  QuestionDataSource,
+  QuestionDescription,
+} from "metabase/query_builder/components/view/ViewHeader/components";
+import { MODAL_TYPES } from "metabase/query_builder/constants";
+import * as Lib from "metabase-lib";
+
+AdHocQuestionLeftSide.propTypes = {
+  question: PropTypes.object.isRequired,
+  originalQuestion: PropTypes.object,
+  isNative: PropTypes.bool,
+  isObjectDetail: PropTypes.bool,
+  isSummarized: PropTypes.bool,
+  onOpenModal: PropTypes.func,
+};
+export function AdHocQuestionLeftSide(props) {
+  const {
+    question,
+    originalQuestion,
+    isNative,
+    isObjectDetail,
+    isSummarized,
+    onOpenModal,
+  } = props;
+
+  const handleTitleClick = () => {
+    const { isEditable } = Lib.queryDisplayInfo(question.query());
+
+    if (isEditable) {
+      onOpenModal(MODAL_TYPES.SAVE);
+    }
+  };
+
+  return (
+    <AdHocLeftSideRoot>
+      <ViewHeaderMainLeftContentContainer>
+        <AdHocViewHeading color="medium">
+          {isNative ? (
+            t`New question`
+          ) : (
+            <QuestionDescription
+              question={question}
+              originalQuestion={originalQuestion}
+              isObjectDetail={isObjectDetail}
+              onClick={handleTitleClick}
+            />
+          )}
+        </AdHocViewHeading>
+      </ViewHeaderMainLeftContentContainer>
+      <ViewHeaderLeftSubHeading>
+        {isSummarized && (
+          <QuestionDataSource
+            className={CS.mb1}
+            question={question}
+            isObjectDetail={isObjectDetail}
+            subHead
+          />
+        )}
+      </ViewHeaderLeftSubHeading>
+    </AdHocLeftSideRoot>
+  );
+}
