+++ frontend/src/metabase/query_builder/components/view/ViewHeader/components/AdHocQuestionLeftSide/index.ts
@@ -1,0 +1,1 @@
+export * from "./AdHocQuestionLeftSide";
