+++ frontend/src/metabase/query_builder/components/view/ViewHeader/components/HeaderCollectionBadge/index.ts
@@ -1,0 +1,1 @@
+export * from "./HeaderCollectionBadge";
