+++ frontend/src/metabase/query_builder/components/view/ViewHeader/components/ViewTitleHeaderRightSide/index.ts
@@ -1,0 +1,1 @@
+export * from "./ViewTitleHeaderRightSide";
