+++ frontend/src/metabase/query_builder/components/view/ViewHeader/components/index.ts
@@ -8,3 +8,8 @@ export * from "./QuestionNotebookButton";
 export * from "./ExploreResultsLink";
 export * from "./FilterHeaderButton";
 export * from "./QuestionSummarizeWidget";
+export * from "./AdHocQuestionLeftSide";
+export * from "./HeaderCollectionBadge";
+export * from "./SavedQuestionLeftSide";
+export * from "./ViewTitleHeaderRightSide";
+export * from "./DashboardBackButton";
